Move failure info clearing for more consistency (#4417)
* move failure info clearing for more consistency

* update null check

* remove null check entirely

* Repro issue with infinite loop if enlistment throws

* Repair absconded test and create new one specifically for MSMQ TransactionScope

* Update all msmq transaction strategies to use the same code pattern for consistency

* Fix msmq strategies to account for manual commit and fix associated failing tests

* reset assert in absconded test

* Clean up formatting

* Update test and class names
